requirements.txt

Flask==1.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.8
flatbuffers==1.12
gast==0.3.3

itsdangerous==1.1.0


Markdown==3.3.3
MarkupSafe==1.1.1
numpy==1.19.5

opencv-python==4.5.1.48

Pillow==8.1.0

requests==2.25.1
requests-oauthlib==1.3.0


